backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
PyJWT==2.8.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
pymysql==1.1.0
web3==7.13.0
python-dotenv==1.0.0
pycryptodome==3.19.0
openai==1.99.9
